foodgram-main/backend/foodgram_backend/api/pagination.py
from django.conf import settings
from rest_framework.pagination import PageNumberPagination


class FoodgramPagination(PageNumberPagination):
    page_size = settings.PAGE_SIZE
    page_size_query_param = 'limit'
    max_page_size = 100
